[H5] Implementation of the create parking service

+++ src/main/java/com/forero/parking/application/port/DbPort.java
@@ -24,5 +24,5 @@ public interface DbPort {
 
     List<ParkingLot> getVehiclesInParking();
 
-    Long save(ParkingLot parkingLot);
+    int save(ParkingLot parkingLot);
 }
